Refactor BDD steps into separate files
