+++ features/step_definitions/context.js
@@ -1,0 +1,9 @@
+module.exports = {
+  browser: null,
+  context: null,
+  page: null,
+  lastShipPos: null,
+  lastPointer: { x: 0, y: 0 },
+  initialAmmo: 0,
+  lastOrbPos: null
+};
