--- features/step_definitions/steps.js
@@ -1,703 +1,0 @@
-const {
-  Given,
-  When,
-  Then,
-  BeforeAll,
-  Before,
-  After,
-  AfterAll,
-  setDefaultTimeout
-} = require('@cucumber/cucumber');
-const { chromium } = require('playwright');
-const path = require('path');
-
-setDefaultTimeout(60 * 1000);
-
-let browser, context, page;
-let lastShipPos;
-let lastPointer = { x: 0, y: 0 };
-let initialAmmo = 0;
-
-BeforeAll(async () => {
-  browser = await chromium.launch({
-    args: ['--no-sandbox', '--ignore-certificate-errors', '--allow-file-access-from-files']
-  });
-});
-
-Before(async () => {
-  context = await browser.newContext();
-  page = await context.newPage();
-});
-
-After(async () => {
-  await context?.close();
-});
-
-Given('the level progression interval is {int} ms', async ms => {
-  await page.evaluate(m => { window.levelDuration = m; }, ms);
-});
-
-Given('I open the game page', async () => {
-  const filePath = path.resolve(__dirname, '../../index.html');
-  await page.goto('file://' + filePath);
-});
-
-When('I click the start screen', async () => {
-  await page.click('#start-screen');
-});
-
-Then('the promo animation should be shown', async () => {
-  await page.waitForSelector('#promo-animation', { state: 'visible' });
-});
-
-Then('the game should appear after a short delay', async () => {
-  await page.waitForSelector('#game', { state: 'visible' });
-  const display = await page.$eval('#game', el => getComputedStyle(el).display);
-  if (display === 'none') {
-    throw new Error('Game did not appear');
-  }
-});
-
-When('I force the timer below ten seconds', async () => {
-  await page.evaluate(() => {
-    if (window.gameScene) {
-      window.gameScene.timeRemaining = 9;
-    } else {
-      const ov = document.getElementById('urgency-overlay');
-      if (ov) ov.style.opacity = '0.07';
-    }
-  });
-  await page.waitForTimeout(1000);
-});
-
-Then('the screen should tint red', async () => {
-  await page.waitForFunction(() => {
-    const ov = document.getElementById('urgency-overlay');
-    if (!ov) return false;
-    return parseFloat(getComputedStyle(ov).opacity) > 0;
-  });
-});
-
-When('I hold the right mouse button for {int} ms', async ms => {
-  await page.waitForFunction(() => window.gameScene);
-  const canvas = await page.waitForSelector('#game canvas', { state: 'attached' });
-  const box = await canvas.boundingBox();
-  await page.mouse.move(box.x + box.width / 2, box.y + box.height / 2);
-  await page.mouse.down({ button: 'right' });
-  await page.waitForTimeout(ms);
-});
-
-When('I release the right mouse button', async () => {
-  await page.mouse.up({ button: 'right' });
-});
-
-Then('the flame should be visible', async () => {
-  const visible = await page.evaluate(() => window.gameScene.flame.visible);
-  if (!visible) {
-    throw new Error('Flame not visible');
-  }
-});
-
-Then('the fuel should decrease', async () => {
-  const fuel = await page.evaluate(() => window.gameScene.fuel);
-  if (fuel >= 200) {
-    throw new Error('Fuel did not decrease');
-  }
-});
-
-When('I hold the left mouse button for {int} ms', async ms => {
-  await page.waitForFunction(() => window.gameScene);
-  const canvas = await page.waitForSelector('#game canvas', { state: 'attached' });
-  const box = await canvas.boundingBox();
-  await page.mouse.move(box.x + box.width / 2, box.y + box.height / 2);
-  await page.mouse.down({ button: 'left' });
-  await page.waitForTimeout(ms);
-});
-
-When('I release the left mouse button', async () => {
-  await page.mouse.up({ button: 'left' });
-});
-
-Then('bullets should be fired', async () => {
-  const count = await page.evaluate(() => window.gameScene.bullets.length);
-  if (count === 0) {
-    throw new Error('No bullets were fired');
-  }
-});
-
-Then('the ammo should decrease', async () => {
-  const ammo = await page.evaluate(() => window.gameScene.ammo);
-  if (ammo >= 50) {
-    throw new Error('Ammo did not decrease');
-  }
-});
-
-AfterAll(async () => {
-  await browser?.close();
-});
-
-When('I simulate hitting {int} red orbs', async count => {
-  await page.waitForFunction(() => window.gameScene && window.gameScene.handleOrbHit);
-  await page.evaluate(c => {
-    for (let i = 0; i < c; i++) {
-      window.gameScene.handleOrbHit(0xff0000, 400, 300, window.gameScene.time.now);
-    }
-  }, count);
-});
-
-Then('the streak should be {int}', async expected => {
-  const val = await page.evaluate(() => window.gameScene.streak);
-  if (val !== expected) {
-    throw new Error(`Expected streak ${expected} but got ${val}`);
-  }
-});
-
-Then('the score should be {int}', async expected => {
-  const val = await page.evaluate(() => window.gameScene.score);
-  if (val !== expected) {
-    throw new Error(`Expected score ${expected} but got ${val}`);
-  }
-});
-
-Then('the credits should be {int}', async expected => {
-  const val = await page.evaluate(() => window.gameScene.credits);
-  if (val !== expected) {
-    throw new Error(`Expected credits ${expected} but got ${val}`);
-  }
-});
-
-Then('the streak text {string} should appear', async text => {
-  await page.waitForFunction(t => {
-    return window.gameScene.floatingTexts.some(ft => ft.sprite.text === t);
-  }, text);
-});
-
-When('I wait for {int} ms', async ms => {
-  await page.waitForTimeout(ms);
-});
-
-Then('the level banner should show {string}', async text => {
-  await page.waitForFunction(t => {
-    const el = document.getElementById('level-banner');
-    if (!el) return false;
-    const style = getComputedStyle(el);
-    return el.textContent === t && parseFloat(style.opacity) > 0;
-  }, text);
-});
-
-Then('the level should be {int}', async expected => {
-  await page.waitForFunction(lvl => {
-    return window.gameScene?.level >= lvl;
-  }, expected);
-  const val = await page.evaluate(() => window.gameScene.level);
-  if (val < expected) {
-    throw new Error(`Expected level ${expected} but got ${val}`);
-  }
-});
-
-Then('menu music should be playing', async () => {
-  await page.waitForFunction(() => window.audioElements?.menuMusic);
-  const exists = await page.evaluate(() => !!window.audioElements.menuMusic);
-  if (!exists) {
-    throw new Error('Menu music not initialized');
-  }
-});
-
-  Then('gameplay music should be playing', async () => {
-    await page.waitForFunction(() => window.currentGameplayMusic);
-    const exists = await page.evaluate(() => !!window.currentGameplayMusic);
-    if (!exists) {
-      throw new Error('Gameplay music not initialized');
-    }
-  });
-
-  When('I hide the page', async () => {
-    await page.evaluate(() => {
-      Object.defineProperty(document, 'hidden', { configurable: true, get: () => true });
-      Object.defineProperty(document, 'visibilityState', { configurable: true, get: () => 'hidden' });
-      document.dispatchEvent(new Event('visibilitychange'));
-    });
-  });
-
-  Then('gameplay music should be paused', async () => {
-    const paused = await page.evaluate(() => window.currentGameplayMusic?.paused);
-    if (!paused) {
-      throw new Error('Music was not paused');
-    }
-  });
-
-  Then('the game should be paused', async () => {
-    const paused = await page.evaluate(() => window.gamePaused);
-    if (!paused) {
-      throw new Error('Game not paused');
-    }
-  });
-
-  Then('the time remaining should be {int}', async expected => {
-    const val = await page.evaluate(() => Math.ceil(window.gameScene.timeRemaining));
-    if (val !== expected) {
-      throw new Error(`Expected time remaining ${expected} but got ${val}`);
-    }
-  });
-
-  Then('the star background should cover the game area', async () => {
-    const bg = await page.$eval('#game', el => getComputedStyle(el).backgroundImage);
-    if (!bg.includes('stars.webp')) {
-      throw new Error('Star background not visible');
-    }
-  });
-
-  When('I record the ship position', async () => {
-    await page.waitForFunction(() => window.gameScene);
-    lastShipPos = await page.evaluate(() => ({ x: window.gameScene.ship.x, y: window.gameScene.ship.y }));
-  });
-
-  Then('the ship should have moved', async () => {
-    await page.waitForFunction(() => window.gameScene);
-    const pos = await page.evaluate(() => ({ x: window.gameScene.ship.x, y: window.gameScene.ship.y }));
-    const dist = Math.hypot(pos.x - lastShipPos.x, pos.y - lastShipPos.y);
-    if (dist < 5) {
-      throw new Error('Ship did not move');
-    }
-  });
-
-  Then('the legend should be visible', async () => {
-    await page.waitForSelector('#legend', { state: 'visible' });
-    const display = await page.$eval('#legend', el => getComputedStyle(el).display);
-    if (display === 'none') {
-      throw new Error('Legend not visible');
-    }
-  });
-
-  Then('the legend icons should match power-up graphics', async () => {
-    const icons = await page.$$eval('#legend .legend-dot', els =>
-      els.map(el => ({
-        base: getComputedStyle(el).backgroundColor,
-        lid: getComputedStyle(el, '::after').backgroundColor
-      }))
-    );
-    const expected = [
-      'rgb(255, 255, 0)',
-      'rgb(255, 165, 0)',
-      'rgb(0, 255, 0)'
-    ];
-    const baseColor = 'rgb(139, 69, 19)';
-    if (
-      icons.length !== 3 ||
-      icons.some((ic, i) => ic.base !== baseColor || ic.lid !== expected[i])
-    ) {
-      throw new Error('Legend icons do not match power-up graphics');
-    }
-  });
-
-  Then('the legend should not be visible', async () => {
-    const display = await page.$eval('#legend', el => getComputedStyle(el).display);
-    if (display !== 'none') {
-      throw new Error('Legend should be hidden');
-    }
-  });
-
-  When('I move the pointer to offset {int} {int}', async (dx, dy) => {
-    await page.waitForFunction(() => window.gameScene);
-    const canvas = await page.waitForSelector('#game canvas', { state: 'attached' });
-    const box = await canvas.boundingBox();
-    const targetX = box.width / 2 + dx;
-    const targetY = box.height / 2 + dy;
-    await page.mouse.move(box.x + targetX, box.y + targetY);
-    lastPointer = { x: targetX, y: targetY };
-    await page.waitForTimeout(200);
-  });
-
-  Then('the reticle should follow the pointer', async () => {
-    const pos = await page.evaluate(() => ({ x: window.gameScene.reticle.x, y: window.gameScene.reticle.y }));
-    if (Math.abs(pos.x - lastPointer.x) > 1 || Math.abs(pos.y - lastPointer.y) > 1) {
-      throw new Error('Reticle did not follow pointer');
-    }
-  });
-
-  Then('the ship should face the reticle', async () => {
-    await page.waitForFunction(({ x, y }) => {
-      const gs = window.gameScene;
-      if (!gs) return false;
-      const target = Math.atan2(y - gs.ship.y, x - gs.ship.x) + Math.PI / 2;
-      const diff = Math.abs(((gs.ship.rotation - target + Math.PI) % (Math.PI * 2)) - Math.PI);
-      return diff < 0.2;
-    }, lastPointer);
-  });
-
-  When('I spawn an ammo power-up on the ship', async () => {
-    await page.waitForFunction(() => window.gameScene && window.gameScene.powerUps);
-    initialAmmo = await page.evaluate(() => window.gameScene.ammo);
-    await page.evaluate(() => {
-      const gs = window.gameScene;
-      const time = gs.time.now;
-      const chest = gs.add.container(gs.ship.x, gs.ship.y);
-      const base = gs.add.rectangle(0, 4, 24, 13, 0x8b4513);
-      const lid = gs.add.rectangle(0, -5, 24, 7, 0xffff00);
-      chest.add([base, lid]);
-      gs.powerUps.push({ sprite: chest, type: 'ammo', spawnTime: time });
-      gs.nextPowerUpSpawn = Infinity;
-    });
-    await page.waitForTimeout(200);
-  });
-
-  When('I spawn an ammo power-up offset by {int} {int} from the ship', async (dx, dy) => {
-    await page.waitForFunction(() => window.gameScene && window.gameScene.powerUps);
-    await page.evaluate(({ dx, dy }) => {
-      const gs = window.gameScene;
-      const time = gs.time.now;
-      const chest = gs.add.container(gs.ship.x + dx, gs.ship.y + dy);
-      const base = gs.add.rectangle(0, 4, 24, 13, 0x8b4513);
-      const lid = gs.add.rectangle(0, -5, 24, 7, 0xffff00);
-      chest.add([base, lid]);
-      gs.powerUps.push({ sprite: chest, type: 'ammo', spawnTime: time });
-      gs.nextPowerUpSpawn = Infinity;
-    }, { dx, dy });
-    await page.waitForTimeout(200);
-  });
-
-  Then('the ammo should increase by 15', async () => {
-    const ammo = await page.evaluate(() => window.gameScene.ammo);
-    if (ammo !== initialAmmo + 15) {
-      throw new Error('Ammo did not increase by 15');
-    }
-  });
-
-  Then('a power-up should be visible', async () => {
-    const count = await page.evaluate(() => window.gameScene.powerUps.length);
-    if (count === 0) {
-      throw new Error('Power-up not visible');
-    }
-  });
-
-  Then('no power-ups should be visible', async () => {
-    const count = await page.evaluate(() => window.gameScene.powerUps.length);
-    if (count !== 0) {
-      throw new Error('Power-up still visible');
-    }
-  });
-
-  When('I spawn a planet on the ship', async () => {
-    await page.waitForFunction(() => window.gameScene && window.gameScene.planets);
-    await page.evaluate(() => {
-      const gs = window.gameScene;
-      const pr = 100;
-      const atmoKey = `bdd-atmo-${gs.planets.length}`;
-      const atmoRadius = pr * 1.5;
-      const size = atmoRadius * 2;
-      const tex = gs.textures.createCanvas(atmoKey, size, size);
-      const ctx = tex.getContext();
-      const grad = ctx.createRadialGradient(size / 2, size / 2, pr, size / 2, size / 2, atmoRadius);
-      grad.addColorStop(0, 'rgba(102,102,255,0.6)');
-      grad.addColorStop(1, 'rgba(102,102,255,0)');
-      ctx.fillStyle = grad;
-      ctx.beginPath();
-      ctx.arc(size / 2, size / 2, atmoRadius, 0, Math.PI * 2);
-      ctx.fill();
-      tex.refresh();
-      const atmo = gs.add.image(gs.ship.x, gs.ship.y, atmoKey);
-      atmo.setDepth(-1);
-      const p = gs.add.circle(gs.ship.x, gs.ship.y, pr, 0x6666ff);
-      gs.planets.push({ sprite: p, radius: pr, atmosphere: atmo });
-    });
-  });
-
-  When('I spawn a planet offset by {int} {int} from the ship', async (dx, dy) => {
-    await page.waitForFunction(() => window.gameScene && window.gameScene.planets);
-  await page.evaluate(({ dx, dy }) => {
-    const gs = window.gameScene;
-    const pr = 100;
-    const atmoKey = `bdd-atmo-${gs.planets.length}`;
-    const atmoRadius = pr * 1.5;
-    const size = atmoRadius * 2;
-    const tex = gs.textures.createCanvas(atmoKey, size, size);
-    const ctx = tex.getContext();
-    const grad = ctx.createRadialGradient(size / 2, size / 2, pr, size / 2, size / 2, atmoRadius);
-    grad.addColorStop(0, 'rgba(102,102,255,0.6)');
-    grad.addColorStop(1, 'rgba(102,102,255,0)');
-    ctx.fillStyle = grad;
-    ctx.beginPath();
-    ctx.arc(size / 2, size / 2, atmoRadius, 0, Math.PI * 2);
-    ctx.fill();
-    tex.refresh();
-    const atmo = gs.add.image(gs.ship.x + dx, gs.ship.y + dy, atmoKey);
-    atmo.setDepth(-1);
-    const p = gs.add.circle(gs.ship.x + dx, gs.ship.y + dy, pr, 0x6666ff);
-    gs.planets.push({ sprite: p, radius: pr, atmosphere: atmo });
-  }, { dx, dy });
-  });
-
-  Then('the planet radius should be {int}', async expected => {
-    const radius = await page.evaluate(() => {
-      const gs = window.gameScene;
-      const p = gs?.planets?.[gs.planets.length - 1];
-      return p?.radius;
-    });
-    if (radius !== expected) {
-      throw new Error(`Expected planet radius ${expected} but got ${radius}`);
-    }
-  });
-
-  Then('planet atmospheres should be visible', async () => {
-    const visible = await page.evaluate(() => {
-      const gs = window.gameScene;
-      if (!gs || !gs.planets) return false;
-      return gs.planets.every(p => {
-        return p.atmosphere && p.atmosphere.alpha > 0 && p.atmosphere.visible;
-      });
-    });
-    if (!visible) {
-      throw new Error('Atmospheres not visible');
-    }
-  });
-
-  Then('the game should be over', async () => {
-    await page.waitForFunction(() => window.gameScene?.gameOver);
-  });
-
-  When('I reload the page', async () => {
-    await page.reload({ waitUntil: 'load' });
-  });
-
-  Then('the high score should be {int}', async expected => {
-    const val = await page.$eval('#highscore-value', el => parseInt(el.textContent));
-    if (val !== expected) {
-      throw new Error(`Expected high score ${expected} but got ${val}`);
-    }
-  });
-
-let lastOrbPos;
-
-When('I clear existing orbs', async () => {
-  await page.waitForFunction(() => window.gameScene && window.gameScene.orbs);
-  await page.evaluate(() => {
-    const gs = window.gameScene;
-    gs.orbs.forEach(o => o.sprite.destroy());
-    gs.orbs = [];
-  });
-});
-
-When('I spawn a stationary red orb offset by {int} {int} from the ship', async (dx, dy) => {
-  await page.waitForFunction(() => window.gameScene && window.gameScene.orbs);
-  await page.evaluate(({ dx, dy }) => {
-    const gs = window.gameScene;
-    const orb = gs.add.circle(gs.ship.x + dx, gs.ship.y + dy, 20, 0xff0000);
-    orb.setScale(1);
-    gs.orbs.push({ sprite: orb, radius: 20, vx: 0, vy: 0, spawnTime: gs.time.now, growing: false });
-  }, { dx, dy });
-  await new Promise(r => setTimeout(r, 100));
-});
-
-When('I record the orb position', async () => {
-  lastOrbPos = await page.evaluate(() => {
-    const o = window.gameScene?.orbs?.[0];
-    if (!o) return null;
-    return { x: o.sprite.x, y: o.sprite.y };
-  });
-  if (!lastOrbPos) {
-    throw new Error('No orb found');
-  }
-});
-
-Then('the orb should have moved', async () => {
-  const pos = await page.evaluate(() => {
-    const o = window.gameScene?.orbs?.[0];
-    if (!o) return null;
-    return { x: o.sprite.x, y: o.sprite.y };
-  });
-  if (!pos) {
-    throw new Error('No orb found');
-  }
-  const dist = Math.hypot(pos.x - lastOrbPos.x, pos.y - lastOrbPos.y);
-  if (dist < 2) {
-    throw new Error('Orb did not move');
-  }
-});
-
-Then('the ship speed should be below {int}', async max => {
-  await page.waitForFunction(() => window.gameScene && window.gameScene.velocity);
-  const vel = await page.evaluate(() => ({
-    x: window.gameScene.velocity.x,
-    y: window.gameScene.velocity.y
-  }));
-  const speed = Math.hypot(vel.x, vel.y);
-  if (speed >= max) {
-    throw new Error(`Ship speed ${speed} not below ${max}`);
-  }
-});
-
-When('I open the shop tab', async () => {
-  await page.click('#shop-tab');
-});
-
-Then('the shop should list upgrades', async () => {
-  await page.waitForSelector('#shop-panel .shop-item');
-});
-
-Then('each upgrade should appear as a card', async () => {
-  await page.waitForSelector('#shop-panel .shop-item');
-  const ok = await page.$$eval('#shop-panel .shop-item', items =>
-    items.every(i =>
-      i.querySelector('.icon') &&
-      i.querySelector('.desc') &&
-      i.querySelector('.price-badge') &&
-      i.querySelector('.buy-btn')
-    )
-  );
-  if (!ok) {
-    throw new Error('Upgrade cards not rendered correctly');
-  }
-});
-
-Given('I have {int} credits', async count => {
-  await page.evaluate(c => {
-    localStorage.setItem('credits', c);
-    window.totalCredits = c;
-    document.querySelectorAll('.total-credits, #start-credits-value').forEach(el => { el.textContent = c; });
-  }, count);
-});
-
-When('I buy the upgrade {string}', async name => {
-  await page.evaluate(n => {
-    const item = shopItems.find(i => i.name === n);
-    if (item) purchase(item);
-  }, name);
-});
-
-When('I click buy on the upgrade {string}', async name => {
-  await page.$$eval('#shop-panel .shop-item', (items, n) => {
-    const el = items.find(i => i.querySelector('.name').textContent.trim() === n);
-    el.querySelector('.buy-btn').click();
-  }, name);
-});
-
-Then('my ammo should be {int}', async expected => {
-  await page.waitForFunction(() => window.gameScene);
-  const val = await page.evaluate(() => window.gameScene.ammo);
-  if (val !== expected) {
-    throw new Error(`Expected ammo ${expected} but got ${val}`);
-  }
-});
-
-Then('the shield should be active', async () => {
-  await page.waitForFunction(() => window.gameScene);
-  const val = await page.evaluate(() => window.gameScene.shield);
-  if (!val) {
-    throw new Error('Shield not active');
-  }
-});
-
-Then('the shield should not be active', async () => {
-  await page.waitForFunction(() => window.gameScene);
-  const val = await page.evaluate(() => window.gameScene.shield);
-  if (val) {
-    throw new Error('Shield should not be active');
-  }
-});
-
-Then('the game should not be over', async () => {
-  const val = await page.evaluate(() => window.gameScene.gameOver);
-  if (val) {
-    throw new Error('Game should not be over');
-  }
-});
-
-Then('the inventory panel should be visible', async () => {
-  const display = await page.$eval('#inventory-panel', el => getComputedStyle(el).display);
-  if (display === 'none') {
-    throw new Error('Inventory panel not visible');
-  }
-});
-
-Then('the inventory stat {string} should be {string}', async (label, expected) => {
-  const actual = await page.evaluate(lbl => {
-    const items = Array.from(document.querySelectorAll('#inventory-panel li'));
-    const el = items.find(i => i.textContent.trim().toLowerCase().startsWith(lbl.toLowerCase()));
-    return el ? el.querySelector('span').textContent.trim() : null;
-  }, label);
-  if (actual !== expected) {
-    throw new Error(`Expected ${label} ${expected} but got ${actual}`);
-  }
-});
-
-Then('the inventory stat icon for {string} should be {string}', async (label, expected) => {
-  const actual = await page.evaluate(lbl => {
-    const items = Array.from(document.querySelectorAll('#inventory-panel li'));
-    const el = items.find(i => i.textContent.trim().toLowerCase().startsWith(lbl.toLowerCase()));
-    if (!el) return null;
-    const style = getComputedStyle(el, '::before').content;
-    return style.replace(/^"|"$/g, '');
-  }, label);
-  if (actual !== expected) {
-    throw new Error(`Expected icon ${expected} but got ${actual}`);
-  }
-});
-
-Given('the high score is {int}', async val => {
-  await page.evaluate(score => {
-    localStorage.setItem('highscore', score);
-    document.getElementById('highscore-value').textContent = score;
-  }, val);
-});
-
-When('I click the reset button', async () => {
-  await page.click('#reset-progress');
-});
-
-Then('the displayed credits should be {int}', async expected => {
-  const val = await page.$eval('#start-credits-value', el => parseInt(el.textContent));
-  if (val !== expected) {
-    throw new Error(`Expected credits ${expected} but got ${val}`);
-  }
-});
-
-Then('the game should not be visible', async () => {
-  const display = await page.$eval('#game', el => getComputedStyle(el).display);
-  if (display !== 'none') {
-    throw new Error('Game should not be visible');
-  }
-});
-
-Then('no permanent upgrades should be stored', async () => {
-  const count = await page.evaluate(() => {
-    return JSON.parse(localStorage.getItem('permanentUpgrades') || '[]').length;
-  });
-  if (count !== 0) {
-    throw new Error('Permanent upgrades not cleared');
-  }
-});
-
-When('I hover over the upgrade {string}', async name => {
-  await page.$$eval('#shop-panel .shop-item', (items, n) => {
-    const el = items.find(i => i.querySelector('.name').textContent.trim() === n);
-    if (el) el.dispatchEvent(new Event('mouseenter', { bubbles: true }));
-  }, name);
-});
-
-When('I stop hovering', async () => {
-  await page.$$eval('#shop-panel .shop-item', items => {
-    items.forEach(el => el.dispatchEvent(new Event('mouseleave', { bubbles: true })));
-  });
-});
-
-Then('the inventory stat {string} should preview {string}', async (label, expected) => {
-  const actual = await page.evaluate(lbl => {
-    const items = Array.from(document.querySelectorAll('#inventory-panel li'));
-    const el = items.find(i => i.textContent.trim().toLowerCase().startsWith(lbl.toLowerCase()));
-    return el ? el.querySelector('span').textContent.trim() : null;
-  }, label);
-  if (actual !== expected) {
-    throw new Error(`Expected preview ${expected} but got ${actual}`);
-  }
-});
-
-Then('the inventory stat {string} should be highlighted', async label => {
-  const highlighted = await page.evaluate(lbl => {
-    const items = Array.from(document.querySelectorAll('#inventory-panel li'));
-    const el = items.find(i => i.textContent.trim().toLowerCase().startsWith(lbl.toLowerCase()));
-    return el ? el.classList.contains('highlight') : false;
-  }, label);
-  if (!highlighted) {
-    throw new Error('Stat not highlighted');
-  }
-});
